fifa_manager/requirements.txt
streamlit
numpy
pandas
plotly_express
seaborn
matplotlib
joblib
